move the types too
