+++ src/exo/api/tests/test_api_error_handling.py
@@ -4,7 +4,7 @@ from typing import Any
 from fastapi import FastAPI, HTTPException
 from fastapi.testclient import TestClient
 
-from ..main import API
+from exo.api.main import API
 
 
 def test_http_exception_handler_formats_openai_style() -> None:
